[MOVE] CategoryModel 및 SubCateogryStatus 모델 이동
